Moved negotiator implementation to mocker module
