+++ packages/http/src/mocker/negotiator/types.ts
